fasilitas kamar

+++ app/Http/Controllers/Admin/AdminKamarController.php
@@ -8,27 +8,38 @@ use App\Models\Kamar;
 use App\Models\KamarImage;
 use App\Models\Kos;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class AdminKamarController extends Controller
 {
-
     public function index(Request $request)
     {
+        $user  = auth::user();
         $kosId = $request->kos_id;
         $sort  = $request->sort;
 
-        // ambil semua kos untuk dropdown
-        $allKos = Kos::orderBy('nama_kos')->get();
+        // Dropdown kos (owner hanya lihat kos miliknya)
+        $allKos = Kos::when($user->role === 'owner', function ($q) use ($user) {
+            $q->where('owner_id', $user->id);
+        })
+            ->orderBy('nama_kos')
+            ->get();
 
-        // query dasar → tampilkan semua kamar dulu
-        $query = Kamar::with(['kos', 'fasilitas']);
+        $query = Kamar::with(['kos', 'fasilitas'])
 
-        // filter kos jika dipilih
+            // FILTER OWNER
+            ->when($user->role === 'owner', function ($q) use ($user) {
+                $q->whereHas('kos', function ($sub) use ($user) {
+                    $sub->where('owner_id', $user->id);
+                });
+            });
+
+        // filter kos dropdown
         if ($kosId) {
             $query->where('kos_id', $kosId);
         }
 
-        // sort kamar
+        // sort
         if ($sort == 'harga_asc') {
             $query->orderBy('harga_bulanan', 'asc');
         } elseif ($sort == 'harga_desc') {
@@ -39,25 +50,30 @@ class AdminKamarController extends Controller
             $query->latest();
         }
 
-        // ambil data
         $items = $query
-        ->latest()
-        ->paginate(9)
-        ->withQueryString();
+            ->paginate(9)
+            ->withQueryString();
 
         return view('admin.kamar.index', compact('items', 'allKos'));
     }
 
-
     public function create()
     {
+        $user = auth::user();
+
         return view('admin.kamar.create', [
-            'kos' => Kos::orderBy('nama_kos')->get(),
+            'kos' => Kos::when($user->role === 'owner', function ($q) use ($user) {
+                $q->where('owner_id', $user->id);
+            })
+                ->orderBy('nama_kos')
+                ->get(),
+
             'fasilitas' => Fasilitas::orderBy('nama_fasilitas')
                 ->get()
                 ->groupBy('kategori'),
         ]);
     }
+
 
 
 
@@ -82,6 +98,14 @@ class AdminKamarController extends Controller
         ]);
 
         // 2️⃣ SIMPAN KAMAR
+        if (auth::user()->role === 'owner') {
+            $kos = Kos::findOrFail($data['kos_id']);
+
+            if ($kos->owner_id !== auth::id()) {
+                abort(403);
+            }
+        }
+
         $kamar = Kamar::create($data);
 
         // 3️⃣ UPLOAD GAMBAR UTAMA
@@ -108,6 +132,13 @@ class AdminKamarController extends Controller
 
     public function edit(Kamar $kamar)
     {
+        if (
+            auth::user()->role === 'owner' &&
+            $kamar->kos->owner_id !== auth::id()
+        ) {
+            abort(403);
+        }
+
         return view('admin.kamar.edit', [
             'kamar' => $kamar->load('fasilitas'),
             'kos'   => Kos::orderBy('nama_kos')->get(),
@@ -121,6 +152,13 @@ class AdminKamarController extends Controller
 
     public function update(Request $request, Kamar $kamar)
     {
+        if (
+            auth::user()->role === 'owner' &&
+            $kamar->kos->owner_id !== auth::id()
+        ) {
+            abort(403);
+        }
+
         $data = $request->validate([
             'kos_id'         => 'required|exists:kos,id',
             'nama_kamar'     => 'required|min:3',
@@ -160,19 +198,34 @@ class AdminKamarController extends Controller
 
     public function destroy(Kamar $kamar)
     {
+        if (
+            auth::user()->role === 'owner' &&
+            $kamar->kos->owner_id !== auth::id()
+        ) {
+            abort(403);
+        }
+
         $kamar->delete();
 
         return back()->with('success', 'Kamar berhasil dihapus');
     }
 
-    public function show(Kamar $kamar)
-{
-    $kamar->load([
-        'kos',
-        'fasilitas',
-        'images'
-    ]);
 
-    return view('admin.kamar.show', compact('kamar'));
-}
+    public function show(Kamar $kamar)
+    {
+        if (
+            Auth::user()->role === 'owner' &&
+            $kamar->kos->owner_id !== auth::id()
+        ) {
+            abort(403);
+        }
+
+        $kamar->load([
+            'kos',
+            'fasilitas',
+            'images'
+        ]);
+
+        return view('admin.kamar.show', compact('kamar'));
+    }
 }
